--- bin/example.json
@@ -1,48 +1,0 @@
-{
-    "public": {
-        "0": {
-            "title": "Tickle",
-            "body": "",
-            "author": "Jeff",
-            "createdAt": 123456789,
-            "updatedAt": 123456789,
-            "likeCount": 0
-        },
-        "1": {
-            "title": "Tackle",
-            "body": "",
-            "author": "Jiff",
-            "createdAt": 123456788,
-            "updatedAt": 123456788,
-            "likeCount": 0
-        }
-    },
-    "private": {
-        "0": {
-            "0":"This is a comment",
-            "1":"This is a comment too"
-            },
-        "1": {
-            "0": "This is a comment for a second poem"
-        }
-    },
-    "user": {
-        "0": {
-            "title": "Tickle",
-            "body": "",
-            "author": "Jeff",
-            "createdAt": 123456789,
-            "updatedAt": 123456789,
-            "pubId":0,
-            "isLive": true
-        },
-        "2": {
-            "title": "This is a poem that isn't live",
-            "body": "",
-            "author": "me",
-            "createdAt": 12345532,
-            "updatedAt": 1233543534,
-            "isLive": false
-        }
-    }
-}
